Update README.md
In this project i was working with my team to diagnose an appropiate statistical model for Randomized Complete Block Design (RCBD) with 3 factors. 

Case obtained from Montgomery's book:
The percentage of hardwood concentration in raw pulp, the freeness, and the cooking time of the pulp are being investigated for their effects on the strength of paper. The data from a three-factor factorial experiment are written in Problem 5-16 page 215 of Montgomery 5th Edition's book.

Using R Programming, we find the perfect model for Randomized Complete Block Design (RCBD) with 3 Factors, hypothesis, descriptive statistics, and multiple comparisons tests with Tukey. The results are:
From the three-factors that have tested, there is a significant interaction effect between the presentation factor and the pressure factor in the paper experiment.